requirements.txt
playwright
google-generativeai
python-dotenv
pyautogui
uiautomation
python-telegram-bot
firebase-admin